Deleting quick action with dismissable

+++ lib/screens/quick_actions_screen/widgets/all_quick_actions_card.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: prefer_const_constructors
+
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:provider/provider.dart';
@@ -8,6 +10,7 @@ import '../../../constants/colors.dart';
 import '../../../constants/sizes.dart';
 import '../../../constants/styles.dart';
 import '../../../constants/types.dart';
+import '../../../providers/transactions_provider.dart';
 import '../../../utils/transactions_utils.dart';
 import '../../add_transaction_screen/add_transaction_screen.dart';
 import '../../../widgets/global/card_action_button.dart';
@@ -19,12 +22,57 @@ class AllQuickActionsCard extends StatelessWidget {
     required this.quickAction,
   }) : super(key: key);
 
-  void deleteQuickAction(BuildContext context) async {
+//* this function will show a dialog to delete and after confirming deleting it will be deleted
+//* or if cancel was clicked the card will come back and won't be deleted
+  Future<bool> deleteQuickAction(BuildContext context) async {
+    bool confirmDelete = false;
+    await showDialog(
+        context: context,
+        builder: (ctx) => AlertDialog(
+              content: Text('Delete Quick Action?'),
+              actions: [
+                TextButton(
+                  onPressed: () {
+                    Navigator.pop(context);
+                  },
+                  child: Text('Cancel'),
+                ),
+                TextButton(
+                  onPressed: () async {
+                    try {
+                      await Provider.of<QuickActionsProvider>(context,
+                              listen: false)
+                          .deleteQuickActions(quickAction.id);
+                    } catch (error) {
+                      showSnackBar(
+                          context, error.toString(), SnackBarType.error);
+                    }
+                    confirmDelete = true;
+                    Navigator.pop(context);
+                  },
+                  child: Text('Delete'),
+                ),
+              ],
+            ));
+    return confirmDelete;
+  }
+
+  //* in this method i will apply the quick action and add the transaction by clicking on the quick action card
+  void applyQuickAction(BuildContext context) async {
+    //? the problem here is that each quick action will have an id , so we can't add the same quick action with the same id to be multiple transactions with the same id
+    //? so i will make the add transaction provider decide the id of the newly added transaction
+
     try {
-      await Provider.of<QuickActionsProvider>(context, listen: false)
-          .deleteQuickActions(quickAction.id);
+      await Provider.of<TransactionProvider>(context, listen: false)
+          .addTransaction(
+        quickAction.title,
+        quickAction.description,
+        quickAction.amount,
+        quickAction.transactionType,
+      );
+      showSnackBar(context, 'Transaction Added', SnackBarType.success, true);
     } catch (error) {
-      showSnackBar(context, error.toString(), SnackBarType.error);
+      showSnackBar(context, error.toString(), SnackBarType.error, true);
     }
   }
 
@@ -32,15 +80,16 @@ class AllQuickActionsCard extends StatelessWidget {
   Widget build(BuildContext context) {
     //* the main container of the card
     return Dismissible(
+      direction: DismissDirection.endToStart,
+      // onDismissed: (direction) => deleteQuickAction(context),
+      confirmDismiss: (direction) => deleteQuickAction(context),
+      background: QuickActionCardBackground(),
       key: Key(quickAction.id),
-      onDismissed: (direction) {},
       child: Container(
+        clipBehavior: Clip.hardEdge,
         width: double.infinity,
         margin: const EdgeInsets.only(bottom: kDefaultPadding / 2),
-        padding: const EdgeInsets.symmetric(
-          horizontal: kDefaultHorizontalPadding / 2,
-          vertical: kDefaultVerticalPadding / 2,
-        ),
+
         decoration: BoxDecoration(
           color: Colors.white,
           borderRadius: BorderRadius.circular(kDefaultBorderRadius),
@@ -49,102 +98,141 @@ class AllQuickActionsCard extends StatelessWidget {
           ],
         ),
         //* the row that hold the main components of the card, ( circular leading icon, title buttons etc...)
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.start,
-          children: [
-            //* transaction type icon
-            Container(
-              alignment: Alignment.center,
-              width: 40,
-              height: 40,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(50),
-                border: Border.all(
-                  width: 2,
-                  color: quickAction.transactionType == TransactionType.income
-                      ? kIncomeColor
-                      : kOutcomeColor,
-                ),
+        child: Material(
+          color: Colors.transparent,
+          child: InkWell(
+            onTap: () => applyQuickAction(context),
+            child: Container(
+              padding: const EdgeInsets.symmetric(
+                horizontal: kDefaultHorizontalPadding / 2,
+                vertical: kDefaultVerticalPadding / 2,
               ),
-              child: Icon(
-                quickAction.transactionType == TransactionType.income
-                    ? Icons.arrow_downward
-                    : Icons.arrow_upward,
-                color: quickAction.transactionType == TransactionType.income
-                    ? kIncomeColor
-                    : kOutcomeColor,
-                size: kDefaultIconSize,
-              ),
-            ),
-            const SizedBox(
-              width: kDefaultPadding / 3,
-            ),
-            //* title and price column
-            Column(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                //* title text widget
-                Text(
-                  quickAction.title,
-                  style: kParagraphTextStyle,
-                ),
-                const SizedBox(
-                  height: kDefaultPadding / 4,
-                ),
-                //* price text widget
-                Text(
-                  '${doubleToString(quickAction.amount)} \$',
-                  style: kSmallTextPrimaryColorStyle,
-                ),
-              ],
-            ),
-            Expanded(
-              //* the container of the actions button
               child: Row(
-                mainAxisAlignment: MainAxisAlignment.end,
+                mainAxisAlignment: MainAxisAlignment.start,
                 children: [
-                  //* for opening the screen to edit the quick action
-                  CardActionButton(
-                    iconData: FontAwesomeIcons.pen,
-                    color: kMainColor,
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (ctx) => AddTransactionScreen(
-                            addTransactionScreenOperations:
-                                AddTransactionScreenOperations.editQuickAction,
-                            editingId: quickAction.id,
-                          ),
-                        ),
-                      );
-                    },
+                  //* transaction type icon
+                  Container(
+                    alignment: Alignment.center,
+                    width: 40,
+                    height: 40,
+                    decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(50),
+                      border: Border.all(
+                        width: 2,
+                        color: quickAction.transactionType ==
+                                TransactionType.income
+                            ? kIncomeColor
+                            : kOutcomeColor,
+                      ),
+                    ),
+                    child: Icon(
+                      quickAction.transactionType == TransactionType.income
+                          ? Icons.arrow_downward
+                          : Icons.arrow_upward,
+                      color:
+                          quickAction.transactionType == TransactionType.income
+                              ? kIncomeColor
+                              : kOutcomeColor,
+                      size: kDefaultIconSize,
+                    ),
                   ),
                   const SizedBox(
-                    width: kDefaultPadding / 4,
+                    width: kDefaultPadding / 3,
                   ),
-                  //* for making a quick action favorite
-                  //* this consumer will provide the data for knowing if it is favorite or not
-                  //* and the toggle favorite
+                  //* title and price column
+                  Column(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      //* title text widget
+                      Text(
+                        quickAction.title,
+                        style: kParagraphTextStyle,
+                      ),
+                      const SizedBox(
+                        height: kDefaultPadding / 4,
+                      ),
+                      //* price text widget
+                      Text(
+                        '${doubleToString(quickAction.amount)} \$',
+                        style: kSmallTextPrimaryColorStyle,
+                      ),
+                    ],
+                  ),
+                  Expanded(
+                    //* the container of the actions button
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.end,
+                      children: [
+                        //* for opening the screen to edit the quick action
+                        CardActionButton(
+                          iconData: FontAwesomeIcons.pen,
+                          color: kMainColor,
+                          onTap: () {
+                            Navigator.push(
+                              context,
+                              MaterialPageRoute(
+                                builder: (ctx) => AddTransactionScreen(
+                                  addTransactionScreenOperations:
+                                      AddTransactionScreenOperations
+                                          .editQuickAction,
+                                  editingId: quickAction.id,
+                                ),
+                              ),
+                            );
+                          },
+                        ),
+                        const SizedBox(
+                          width: kDefaultPadding / 4,
+                        ),
+                        //* for making a quick action favorite
+                        //* this consumer will provide the data for knowing if it is favorite or not
+                        //* and the toggle favorite
 
-                  CardActionButton(
-                    iconData: quickAction.isFavorite
-                        ? FontAwesomeIcons.solidHeart
-                        : FontAwesomeIcons.heart,
-                    color: kDeleteColor,
-                    //* this will toggle the favorite for each quick action
-                    onTap: () async {
-                      await Provider.of<QuickActionsProvider>(context,
-                              listen: false)
-                          .toggleFavouriteQuickAction(quickAction.id);
-                    },
-                  ),
+                        CardActionButton(
+                          iconData: quickAction.isFavorite
+                              ? FontAwesomeIcons.solidHeart
+                              : FontAwesomeIcons.heart,
+                          color: kDeleteColor,
+                          //* this will toggle the favorite for each quick action
+                          onTap: () async {
+                            await Provider.of<QuickActionsProvider>(context,
+                                    listen: false)
+                                .toggleFavouriteQuickAction(quickAction.id);
+                          },
+                        ),
+                      ],
+                    ),
+                  )
                 ],
               ),
-            )
-          ],
+            ),
+          ),
         ),
+      ),
+    );
+  }
+}
+
+class QuickActionCardBackground extends StatelessWidget {
+  const QuickActionCardBackground({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      alignment: Alignment.centerRight,
+      padding: EdgeInsets.only(right: kDefaultPadding / 2),
+      margin: const EdgeInsets.only(bottom: kDefaultPadding / 2),
+      decoration: BoxDecoration(
+        color: Colors.red,
+        borderRadius: BorderRadius.circular(kDefaultBorderRadius),
+      ),
+      child: Icon(
+        Icons.delete,
+        color: Colors.white,
+        size: kDefaultIconSize,
       ),
     );
   }
